refactor: put models in /models
