+++ src/app/models/class.ts
