+++ src/app/models/diagram.ts
